+++ auto_route/lib/auto_route.dart
@@ -9,7 +9,6 @@ export 'src/router/auto_route_page.dart';
 export 'src/router/auto_router_x.dart';
 export 'src/router/controller/controller_scope.dart';
 export 'src/router/controller/routing_controller.dart';
-export 'src/router/delegate/auto_router_delegate.dart';
 export 'src/router/parser/route_information_parser.dart';
 export 'src/router/provider/auto_route_information_provider.dart';
 export 'src/router/widgets/auto_back_button.dart';