+++ auto_route_generator/lib/src/resolvers/route_config_resolver.dart
@@ -200,6 +200,10 @@ class RouteConfigResolver {
     var replacementInRouteName = _routerConfig.replaceInRouteName;
 
     final constructor = classElement.unnamedConstructor;
+    throwIf(
+      constructor == null,
+      'Route widgets must have an unnamed constructor',
+    );
     var hasConstConstructor = false;
     var params = constructor!.parameters;
     var parameters = <ParamConfig>[];