+++ auto_route/CHANGELOG.md
@@ -1,4 +1,14 @@
 # ChangeLog
+## [3.1.0]
+- Prefer previous-current route index as new current index when updating tab routes #797
+- Improve native navigation history
+- Use browser history as navigation history in web
+- Fix path/query params not updating issue #809
+- Fix generic nullable types are generated as non-nullables #811
+- Override toString method inside of generated arg classes [FR] #820
+- Support redirect paths with path params [FR] #818
+- Fix replace/replaceNamed/replaceAll do not replace current url in browser #781
+
 ## [3.0.4]
 - Fix IndexedStack widgets are not updated on tabs routes change.
 - Fix RangeError when updating routes in AutoTabsScaffold #788